requirements.txt
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0.1
scikit-learn>=1.5.0
matplotlib>=3.8.0
torch>=2.2.0
mlflow>=2.20.0
jupyter>=1.0.0
notebook>=7.0.0
requests>=2.32.0
datasets>=2.20.0
transformers>=4.44.0
optuna>=3.6.1
pyarrow>=16.1.0
tqdm>=4.66.0

fastapi
uvicorn
pydantic
python-multipart
